# hosted-agent-kit/tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AgentAdminView": {
        "properties": {
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "agent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Version new sessions are pinned to; null means the latest.",
            "title": "Agent Version"
          },
          "circuit_state": {
            "description": "closed, half_open, open or disabled.",
            "title": "Circuit State",
            "type": "string"
          },
          "conditions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ConditionView"
            },
            "title": "Conditions",
            "type": "array"
          },
          "generation": {
            "default": 1,
            "description": "Changes when the pool's configuration changes.",
            "title": "Generation",
            "type": "integer"
          },
          "max_sessions": {
            "title": "Max Sessions",
            "type": "integer"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "observed_generation": {
            "default": 0,
            "description": "The generation the controllers have applied. Behind means pending.",
            "title": "Observed Generation",
            "type": "integer"
          },
          "queue_depth": {
            "title": "Queue Depth",
            "type": "integer"
          },
          "queue_max_depth": {
            "title": "Queue Max Depth",
            "type": "integer"
          },
          "reserved_slots": {
            "title": "Reserved Slots",
            "type": "integer"
          },
          "sessions_available": {
            "title": "Sessions Available",
            "type": "integer"
          },
          "sessions_leased": {
            "title": "Sessions Leased",
            "type": "integer"
          },
          "sessions_retiring": {
            "title": "Sessions Retiring",
            "type": "integer"
          },
          "sessions_total": {
            "title": "Sessions Total",
            "type": "integer"
          },
          "sessions_unavailable": {
            "title": "Sessions Unavailable",
            "type": "integer"
          },
          "telemetry_enabled": {
            "title": "Telemetry Enabled",
            "type": "boolean"
          },
          "version_drain": {
            "default": "never",
            "description": "never, unbound or idle: which idle old-version sessions go.",
            "title": "Version Drain",
            "type": "string"
          }
        },
        "required": [
          "agent_name",
          "mode",
          "max_sessions",
          "sessions_total",
          "sessions_available",
          "sessions_leased",
          "sessions_unavailable",
          "sessions_retiring",
          "reserved_slots",
          "queue_depth",
          "queue_max_depth",
          "telemetry_enabled",
          "circuit_state"
        ],
        "title": "AgentAdminView",
        "type": "object"
      },
...
      "DeleteAccepted": {
        "properties": {
          "detail": {
            "default": "The session is in use and will be deleted when its lease ends.",
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "default": "deferred",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "DeleteAccepted",
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "phase": {
            "description": "Where the request failed: request, auth, queue, create, invoke or stream.",
            "title": "Phase",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          },
          "retry_safe": {
            "description": "True when retrying the same request cannot repeat work the agent already did.",
            "title": "Retry Safe",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "error_code",
          "phase",
          "retry_safe"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token. The user identity is the `oid` claim.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "BFF and worker-pool controller for Microsoft Foundry hosted agents.",
    "title": "Foundry Hosted Agent Pool",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/agents": {
      "get": {
        "operationId": "list_agents_v1_admin_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentAdminView"
                  },
                  "title": "Response List Agents V1 Admin Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Agents",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/agents/{agent_name}": {
      "get": {
        "operationId": "get_agent_v1_admin_agents__agent_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Agent Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolResourceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/agents/{agent_name}/sessions": {
      "get": {
        "operationId": "list_sessions_v1_admin_agents__agent_name__sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Agent Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionAdminView"
                  },
                  "title": "Response List Sessions V1 Admin Agents  Agent Name  Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/agents/{agent_name}/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session_v1_admin_agents__agent_name__sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Agent Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,256}$",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Session",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "get_session_v1_admin_agents__agent_name__sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Agent Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,256}$",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAdminView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/agents/{agent_name}/sync": {
      "post": {
        "operationId": "sync_agent_v1_admin_agents__agent_name__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Agent Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncReportView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sync Agent",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/config/reload": {
      "post": {
        "description": "Re-read the agent pool document and apply it. Adding or removing agents needs a restart.",
        "operationId": "reload_config_v1_admin_config_reload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReloadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reload Config",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/events": {
      "get": {
        "description": "What the controllers did and why, newest first. Not state: nothing reads it.",
        "operationId": "list_events_v1_admin_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Name"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventView"
                  },
                  "title": "Response List Events V1 Admin Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}